requirements.txt
# Core PDF and Image Processing
pymupdf>=1.23.0
pdfplumber>=0.10.0
Pillow>=10.0.0

# Presentation and Video
python-pptx>=0.6.23
moviepy>=1.0.3

# Text-to-Speech
gTTS>=2.5.0
pyttsx3>=2.90
pydub>=0.25.1

# AI and NLP (compatible versions)
transformers>=4.30.0
torch>=2.0.0
sentencepiece>=0.1.99

# Compatible NumPy version
numpy>=1.21.0,<2.0.0

# Utilities
tqdm>=4.65.0
soundfile>=0.12.1

# Optional UI
streamlit>=1.28.0